requirements.txt
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
requests
tqdm
python-dotenv
